+++ augur/export_v2.py
@@ -182,7 +182,7 @@ def update_deprecated_names(name):
 def get_values_across_nodes(node_attrs, key):
     vals = set()
     for data in node_attrs.values():
-        if data.get(key):
+        if is_valid(data.get(key)):
             vals.add(data.get(key))
     return vals